feat: auto revealing sticky lower nav bar
